requirements.txt
requests
lxml
huggingface_hub
tqdm
python-dotenv